refactor(generators): extract UI catalog into separate module
Split UI catalog types and algorithm definitions out of generators/mod.rs:
- catalog.rs: AlgoControlKindV1, AlgoControlPresentationV1, AlgoControlV1,
  AlgoDefinitionV1, AlgoUiEntryV1, shared control constants, ALGO_DEFINITIONS_V1,
  algo_definitions_v1(), algo_ui_catalog_v1()
- mod.rs: re-exports catalog types for backward compatibility, retains engine
  core (LineRenderConfig, AlgoRuntimeState, warp_phase, render_algo_sample)
